split up imaging module
